\name{print.QCLscan}
\alias{print.QCLscan}

\title{Print the results of a QCL genome scans}

\description{
  Print the results of a \code{\link{QCLscan}}
}

\usage{
\method{print}{QCLscan}(x, \dots)
}

\arguments{
 \item{x}{ An object output by the function \code{\link{QCLscan}}.}
 \item{\dots}{ Ignored.} 
}

\details{
  None.
}

\value{
  None. 
}

\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}\cr
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}

\examples{
  #TODO
}

\seealso{
  \itemize{
    \item \code{\link{QCLscan}} - Scan for QCL
    \item \code{\link{plot.QCLscan}} - Plot a QCLscan object
	}
}

\references{
  TODO
}

\keyword{print}
