\name{qcl-internal}
\alias{QCLscan.internal}
\alias{correlationmatrix}
\alias{summary.QCL}
\alias{count.QCL}
\alias{countVThreshold}

\alias{read.QCL.permute}
\alias{significance.QCL}
\alias{significance.QCLHotSpot}

\alias{plot.QCLscan}
\alias{image.QCLscan}

\alias{plot.markerQCL}
\alias{plot.markerQCL2}
\alias{plot.markerQCL.detailed}

\alias{lodscorestoscanone}
\alias{gcLoop}
\alias{test.QCL}

\title{Internal QCL functions}

\description{
  Internal QCL functions. These are generally not to be called by the
  user.
}

\author{Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
