\name{qcl-internal}

\alias{correlationmatrix}
\alias{summary.QCL}
\alias{count.QCL}

\alias{read.QCLpermute}
\alias{significance.QCL}
\alias{significance.QCLHotSpot}

\alias{plot.markerQCL}
\alias{plot.markerQCL2}
\alias{plot.markerQCL.detailed}

\alias{getCorrelatedPhenotypes}
\alias{lodscorestoscanone}
\alias{gcLoop}

\alias{getMarkerOffsets}

\alias{write.nodeAttributeFile}

\alias{print.QCLpermute}

\alias{KEGG.pathway.parser}
\alias{nodesToGenes}

\alias{has_d}
\alias{has_snow}
\alias{has_rcurl}
\alias{has_rqtl}
\alias{get_QclEnv}

\alias{QCLpermute.internal}
\alias{QCLpermuteMC}
\alias{QCLscantoScanone}
\alias{QCLscoretoPvalue}
\alias{QCLtoLOD}
\alias{QCLtoLODvector}
\alias{QCLtoPvalue}
\alias{QCLtoPvalue.internal}
\alias{QTLscan}
\alias{extrapolateBeyondRange}
\alias{getPermuteThresholds}
\alias{getRVM}
\alias{hist.QCLpermute}
\alias{plot.QCLpermute}
\alias{plotAsLOD}
\alias{plotAsStackedHist}

\alias{matrix_correlation}

\title{Internal QCL functions}

\description{
  Internal QCL functions. These are generally not to be called by the
  user.
}

\author{Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
