\name{qcl-internal}
\alias{QCLscan.internal}
\alias{correlationmatrix}
\alias{summary.QCL}
\alias{count.QCL}
\alias{countVThreshold}

\alias{read.QCLpermute}
\alias{significance.QCL}
\alias{significance.QCLHotSpot}

\alias{plot.QCLscan}
\alias{image.QCLscan}

\alias{plot.markerQCL}
\alias{plot.markerQCL2}
\alias{plot.markerQCL.detailed}

\alias{lodscorestoscanone}
\alias{gcLoop}
\alias{test.QCL}

\alias{QCLscan.deprecated}
\alias{QCLscan.internal.deprecated}
\alias{QCLscan.internal}
\alias{QCLscanToSIF}
\alias{bplot.QCLscan}
\alias{getCorrelatedPhenotypes}
\alias{getMarkerOffsets}

\alias{write.nodeAttributeFile}

\alias{print.QCLscan}
\alias{print.QCLpermute}

\alias{KEGG.pathway.parser}
\alias{nodesToGenes}

\alias{correlation}
\alias{matrix_correlation}

\title{Internal QCL functions}

\description{
  Internal QCL functions. These are generally not to be called by the
  user.
}

\author{Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
