\name{qcl-package}
\alias{qcl}
\alias{QCL}
\docType{package}
\title{
	qcl - QCL mapping in experimental crosses
}
\description{
	Includes the following functionality:
	\itemize{
    \item \code{\link{QCL.scan}} - Scan for QCL
    \item \code{\link{QCL.permute}} - Permutation routine for QCL mapping
	}
}
\details{
	More detailed information and/or examples are given per function as needed.
}
\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\references{
	\itemize{
    \item TODO
	}
}
\keyword{package}

\seealso{
	\itemize{
    \item \code{\link{QCL.scan}} - Scan for QCL
    \item \code{\link{QCL.permute}} - Permutation routine for QCL mapping
	}
}
